Reorganize CredRank-related modules
This commit moves the core CredRank modules into a folder called
core/credrank. The motivation is that I want to start splitting apart
MarkovProcessGraph into smaller modules, and doing so without moving the
files would result in a disorganized core directory.
